requirements.txt
aiofiles==23.2.1
aiohttp==3.8.4
aiosignal==1.3.1
altair==5.1.1
annotated-types==0.5.0
anyio==3.7.1
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
black==23.7.0
blis==0.7.10
bs4==0.0.1
catalogue==2.0.9
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.4
comm==0.1.3
confection==0.1.1
contourpy==1.1.0
cycler==0.11.0
cymem==2.0.7
dataclasses-json==0.5.9
debugpy==1.6.7
decorator==5.1.1
discord.py==2.3.1
dnspython==2.3.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.6.0/en_core_web_sm-3.6.0-py3-none-any.whl
exceptiongroup==1.1.3
executing==1.2.0
fastapi==0.103.1
ffmpy==0.3.1
filelock==3.12.3
fonttools==4.42.1
frozenlist==1.3.3
fsspec==2023.9.0
gradio==3.42.0
gradio_client==0.5.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
huggingface-hub==0.16.4
idna==3.4
importlib-resources==6.0.1
ipykernel==6.24.0
ipython==8.14.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.1
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter_client==8.3.0
jupyter_core==5.3.1
kiwisolver==1.4.5
langchain==0.0.273
langcodes==3.3.0
langsmith==0.0.26
loguru==0.7.0
lxml==4.9.3
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.2
matplotlib-inline==0.1.6
mmh3==4.0.0
multidict==6.0.4
murmurhash==1.0.9
mypy==1.5.1
mypy-extensions==1.0.0
nest-asyncio==1.5.6
nltk==3.8.1
numexpr==2.8.4
numpy==1.25.1
openai==0.28.0
openapi-schema-pydantic==1.2.4
orjson==3.9.5
packaging==23.1
pandas==2.1.0
parso==0.8.3
pathspec==0.11.1
pathy==0.10.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
pinecone-client==2.2.2
platformdirs==3.9.1
preshed==3.0.8
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.4.2
pydantic_core==2.10.1
pydub==0.25.1
Pygments==2.15.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0
pyzmq==25.1.0
referencing==0.30.2
regex==2023.6.3
requests==2.31.0
rpds-py==0.10.2
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
semantic-version==2.10.0
six==1.16.0
slack-bolt==1.18.0
slack-sdk==3.21.3
smart-open==6.3.0
sniffio==1.3.0
soupsieve==2.4.1
spacy==3.6.1
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SQLAlchemy==2.0.18
srsly==2.4.7
stack-data==0.6.2
starlette==0.27.0
tenacity==8.2.2
thinc==8.1.12
tiktoken==0.4.0
tokenize-rt==5.1.0
tomli==2.0.1
toolz==0.12.0
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.3
uvicorn==0.23.2
wasabi==1.1.2
wcwidth==0.2.6
websockets==11.0.3
wget==3.2
yarl==1.9.2
